Day24 컴포넌트 렌더링 및 수정페이지
