+++ homework/03-day21/src/app/boards/styles.module.css
@@ -85,6 +85,7 @@
 
     border-radius: 0.5rem;
     border: 1px solid #F2F2F2;
+    cursor: pointer;
 }
