+++ homework/03-day21/src/components/boards-write/styles.module.css
@@ -143,6 +143,7 @@
     border: 1px solid #000;
     background: #fff;
   
+    color: #000;
     font-family: 'Pretendard Variable', sans-serif;
     font-size: 1.125rem;
     font-style: normal;
